review
calloc instead of malloc.

free `stags` and `mtags` since they are initialized with calloc.

should not use `mtev_dyn_buffer_size` because the return is double the actul size we intent to use. use the value used to ensure the size instead.

should not use `mtev_dyn_buffer_used` when modifying the buffer through noit functions instead of buffer functions because return will be inacurate.

use `size_t` insted of `int32_t`.

use `mtev_dyn_buffer_data` instead of creating new pointer.

use const.

clean up tests and assert the expected tag pair sizes.

+++ src/noit_metric.h
@@ -105,7 +105,7 @@ typedef enum {
 #define NOIT_TAG_MAX_CAT_LEN  254
 #define NOIT_TAG_DECODED_SEPARATOR 0x1f
 
-static const int32_t NOIT_IMPLICIT_TAG_MAX_PAIR_LEN =
+static const size_t NOIT_IMPLICIT_TAG_MAX_PAIR_LEN =
     MAX_METRIC_TAGGED_NAME + sizeof("__name:") - 1;
 
 typedef struct {
